requirements.txt
requests>=2.28,<3
pyyaml>=6.0
beautifulsoup4>=4.12
python-dateutil>=2.8
